scenes are accessible as methods instead of array
It seems like a better idea to store scenes as methods in the stage
rather than in a collection. This is because the stage probably
shouldn't be interacted with as a state machine that transitions to the
next scene until the end. There are cases where it makes sense to jump
between stages until the end of the test.

I'm also not sure how it makes sense to constantly call +next_scene+
throughout the test framework (Cucumber, RSpec, etc...)