store stage information in a narrator
This is a large change that is largely influenced by the StateMachine
project. The two big improvements are allowing stages to be inherited
from each other and improving the testability of creating actors and
stages.

Narrator is an object that acts like a singleton over each Stage class.
It contains all the actors and stages which the stage object delegates
to through method_missing.